+++ src/main/java/com/plexter/sellter/agent/product/service/ProductService.java
@@ -6,31 +6,35 @@ import com.plexter.sellter.agent.product.vo.SltrLcProdDtlVO;
 import com.plexter.sellter.agent.product.vo.SltrLcProdRegistVO;
 import com.plexter.sellter.agent.product.vo.SltrLcProdVO;
 import com.plexter.sellter.agent.util.ParsingCommonUtil;
-import org.apache.ibatis.annotations.Mapper;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class ProductService {
 
-    @Mapper
+    @Autowired
     ProductMapper mapper;
 
     // Create
     public String saveSltrLcProdDtlImg(SltrLcProdDtlImgVO vo){
         vo.setOBJ_ID(ParsingCommonUtil.generateObjId("LCPRDDTIMG"));
-        return mapper.saveSltrLcProdDtlImg(vo);
+        mapper.saveSltrLcProdDtlImg(vo);
+        return vo.getOBJ_ID();
     }
     public String saveSltrLcProdDtl(SltrLcProdDtlVO vo){
         vo.setOBJ_ID(ParsingCommonUtil.generateObjId("LCPRDDTL"));
-        return mapper.saveSltrLcProdDtl(vo);
+        mapper.saveSltrLcProdDtl(vo);
+        return vo.getOBJ_ID();
     }
     public String saveSltrLcProdRegist(SltrLcProdRegistVO vo){
         vo.setOBJ_ID(ParsingCommonUtil.generateObjId("LCPRDREG"));
-        return mapper.saveSltrLcProdRegist(vo);
+        mapper.saveSltrLcProdRegist(vo);
+        return vo.getOBJ_ID();
     }
     public String saveSltrLcProd(SltrLcProdVO vo){
         vo.setOBJ_ID(ParsingCommonUtil.generateObjId("LCPRD"));
-        return mapper.saveSltrLcProd(vo);
+        mapper.saveSltrLcProd(vo);
+        return vo.getOBJ_ID();
     }
 
 
@@ -39,40 +43,79 @@ public class ProductService {
         if(vo.getOBJ_ID().isEmpty()){
             throw new NullPointerException("Object id is empty");
         }
-        return mapper.updateSltrLcProdDtlImg(vo);
+        mapper.updateSltrLcProdDtlImg(vo);
+        return mapper.getSltrLcProdDtlImgById(vo.getOBJ_ID());
     }
     public SltrLcProdDtlVO updateSltrLcProdDtl(SltrLcProdDtlVO vo){
         if(vo.getOBJ_ID().isEmpty()){
             throw new NullPointerException("Object id is empty");
         }
-        return mapper.updateSltrLcProdDtl(vo);
+        mapper.updateSltrLcProdDtl(vo);
+        return mapper.getSltrLcProdDtlById(vo.getOBJ_ID());
     }
     public SltrLcProdRegistVO updateSltrLcProdRegist(SltrLcProdRegistVO vo){
         if(vo.getOBJ_ID().isEmpty()){
             throw new NullPointerException("Object id is empty");
         }
-        return mapper.updateSltrLcProdRegist(vo);
+        mapper.updateSltrLcProdRegist(vo);
+        return mapper.getSltrLcProdRegistById(vo.getOBJ_ID());
     }
     public SltrLcProdVO updateSltrLcProd(SltrLcProdVO vo){
         if(vo.getOBJ_ID().isEmpty()){
             throw new NullPointerException("Object id is empty");
         }
-        return mapper.updateSltrLcProd(vo);
+        mapper.updateSltrLcProd(vo);
+        return mapper.getSltrLcProdById(vo.getOBJ_ID());
     }
 
 
     // Delete By Id
     public SltrLcProdDtlImgVO deleteSltrLcProdDtlImgById(String id){
-        return mapper.deleteSltrLcProdDtlImgById(id);
+        SltrLcProdDtlImgVO vo = this.getSltrLcProdDtlImgById(id);
+        if(vo == null){
+            throw new NullPointerException("Id is not defined" + id);
+        }
+        mapper.deleteSltrLcProdDtlImgById(id);
+        return vo;
     }
     public SltrLcProdDtlVO deleteSltrLcProdDtlById(String id){
-        return mapper.deleteSltrLcProdDtlById(id);
+        SltrLcProdDtlVO vo = this.getSltrLcProdDtlById(id);
+        if(vo == null){
+            throw new NullPointerException("Id is not defined" + id);
+        }
+        mapper.deleteSltrLcProdDtlById(id);
+        return vo;
     }
     public SltrLcProdRegistVO deleteSltrLcProdRegistById(String id){
-        return mapper.deleteSltrLcProdRegistById(id);
+        SltrLcProdRegistVO vo = this.getSltrLcProdRegistById(id);
+        if(vo == null){
+            throw new NullPointerException("Id is not defined" + id);
+        }
+        mapper.deleteSltrLcProdRegistById(id);
+        return vo;
     }
     public SltrLcProdVO deleteSltrLcProdById(String id){
-        return mapper.deleteSltrLcProdById(id);
+        SltrLcProdVO vo = this.getSltrLcProdById(id);
+        if(vo == null){
+            throw new NullPointerException("Id is not defined" + id);
+        }
+        mapper.deleteSltrLcProdById(id);
+        return vo;
+    }
+
+
+    // Get By Id
+    public SltrLcProdDtlImgVO getSltrLcProdDtlImgById(String id){
+        return mapper.getSltrLcProdDtlImgById(id);
+    }
+    public SltrLcProdDtlVO getSltrLcProdDtlById(String id){
+        return mapper.getSltrLcProdDtlById(id);
+    }
+    public SltrLcProdRegistVO getSltrLcProdRegistById(String id){
+        return mapper.getSltrLcProdRegistById(id);
+    }
+    public SltrLcProdVO getSltrLcProdById(String id){
+        return mapper.getSltrLcProdById(id);
     }
 
 }
